+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New KiCost variant style.
 - `skip_if_no_field` and `invert` options to the regex used in the generic
   filter.
+- Board view file format export (#69)
 - Basic KiCost support (experimental).
 - Basic internal BoM and KiCost integration (experimental)
 - Experimental mechanism to change 3D models according to the variant.